requirements.txt
pyrogram
tgcrypto
motor
marshmallow
umongo
requests
psutil
pymongo[srv]
python-dotenv
pytz
requests
